projects/chungnam-weather-tour/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
requests==2.32.5
python-dotenv==1.2.2
jinja2==3.1.6
pydantic==2.13.0
